chore: Test ExpectSuite under Scala 2.13
