+++ build.sbt
@@ -54,11 +54,13 @@ commands += Command.command("download-scala-library") { s =>
   s
 }
 commands += Command.command("save-expect") { s =>
-  s"++$LatestScala212" ::
-    "semanticdbScalacPlugin/compile" ::
-    "semanticdbIntegration/clean" ::
-    "semanticdbIntegration/compile" ::
-    "testsJVM/Test/runMain scala.meta.tests.semanticdb.SaveExpectTest" :: s
+  List(LatestScala212, LatestScala213).foldLeft(s) { case (s, version) =>
+    s"++$version" ::
+      "semanticdbScalacPlugin/compile" ::
+      "semanticdbIntegration/clean" ::
+      "semanticdbIntegration/compile" ::
+      "testsJVM/Test/runMain scala.meta.tests.semanticdb.SaveExpectTest" :: s
+  }
 }
 commands += Command.command("save-manifest") { s =>
   "testsJVM/test:runMain scala.meta.tests.semanticdb.SaveManifestTest" :: s
@@ -417,6 +419,7 @@ lazy val testSettings: List[Def.SettingsDefinition] = List(
   buildInfoKeys := Seq[BuildInfoKey](
     scalaVersion,
     "latestScala212Version" -> LatestScala212,
+    "latestScala213Version" -> LatestScala213,
     "databaseSourcepath" ->
       (ThisBuild / baseDirectory).value.getAbsolutePath,
     "commonJVMClassDirectory" -> (common.jvm / Compile / classDirectory).value.getAbsolutePath,
