+++ tests/jvm/src/test/scala/scala/meta/tests/semanticdb/ScalaVersion.scala
@@ -12,24 +12,13 @@ object ScalaVersion {
   case object Scala213 extends Version
   case class Full(ver: String) extends Version
 
-  // both the compiler and stdlib are different between Scala versions.
-  // For the sake of simplicity, we only run the expect test against the
-  // output of 2.12. It's possible to add another expect file for 2.11
-  // later down the road if that turns out to be useful.
-  def isLatest212: Boolean = {
-    version == BuildInfo.latestScala212Version
+  def isLatest212Or213: Boolean = {
+    version == BuildInfo.latestScala212Version ||
+    version == BuildInfo.latestScala213Version
   }
 
   def is212: Boolean = {
     version.startsWith("2.12")
-  }
-
-  def doIfLatest212[T](what: String)(thunk: => T): Unit = {
-    if (isLatest212) {
-      thunk
-    } else {
-      println(s"Skipping $what because scalaVersion is ${Properties.versionNumberString}, which is not the latest Scala 2.12 version ${BuildInfo.latestScala212Version}")
-    }
   }
 
   def doIf212[T](what: String)(thunk: => T): Unit = {
